Separate out the Name API
This commit separates out the name generation logic from the rest
of the design.

The basic `Rules` datatype is now the simple:

```haskell
data
  = Rules
    { rules :: inputs -> SmallIO ( outputs, Map RuleId Rule )
    , actions :: inputs -> Map ActionId Action
    }
```

The API module then defines an monadic-style interface with the type

  type RulesM = inputs -> ActionsM ( RulesT SmallIO outputs )

which supports the `registerRule` and `registerAction` to ensure that
`RuleId`s and `ActionId`s are correct by construction.